requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
cloudinary==1.36.0
dj-database-url==2.1.0
Django==4.2.5
django-cloudinary-storage==0.3.0
django-filter==23.3
djangorestframework==3.14.0
idna==3.4
Pillow==10.0.1
psycopg2-binary==2.9.8
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.6
